Some code reorganization
