+++ src/filesystem.cpp
@@ -463,3 +463,12 @@ bool can_write_to_dir( const std::string &dir_path )
 
     return remove_file( dummy_file );
 }
+
+std::string get_pid_string()
+{
+#if defined _WIN32
+    return to_string( GetCurrentProcessId() );
+#else
+    return to_string( getpid() );
+#endif
+}
